generation/requirements.txt
mlflow
pytorch-lightning
pyyaml
torchvision
pandas
numpy
rasterio
scikit-image
pytorch-msssim
diffusers